test(e2e): restructure resize suite to house pattern, pin kai config
Follow the DescribeAllocateElasticSpecs shape: Ordered container, queues and
skips once in BeforeAll, TestContextCleanup in AfterEach, ClusterCleanup in
AfterAll. Add feature_flags.SetInPlacePodResizeValidation to patch the kai
config's inPlacePodResize settings (with admission rollout wait) and pin
validateQuota=true for the suite, restoring defaults in AfterAll.
