+++ test/e2e/suites/resize/resize_suite_test.go
@@ -13,6 +13,8 @@ import (
 	. "github.com/onsi/gomega"
 )
 
+var _ = DescribeResizeSpecs()
+
 func TestInPlacePodResize(t *testing.T) {
 	utils.SetLogger()
 	RegisterFailHandler(Fail)
