--- test/e2e/suites/resize/resize_test.go
@@ -1,183 +1,0 @@
-/*
-Copyright 2026 NVIDIA CORPORATION
-SPDX-License-Identifier: Apache-2.0
-*/
-package resize
-
-import (
-	"context"
-	"strconv"
-	"strings"
-	"time"
-
-	. "github.com/onsi/ginkgo/v2"
-	. "github.com/onsi/gomega"
-	v1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/api/resource"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/types"
-	"k8s.io/client-go/util/retry"
-
-	v2 "github.com/kai-scheduler/KAI-scheduler/pkg/apis/scheduling/v2"
-	testcontext "github.com/kai-scheduler/KAI-scheduler/test/e2e/modules/context"
-	"github.com/kai-scheduler/KAI-scheduler/test/e2e/modules/resources/capacity"
-	"github.com/kai-scheduler/KAI-scheduler/test/e2e/modules/resources/rd"
-	"github.com/kai-scheduler/KAI-scheduler/test/e2e/modules/resources/rd/queue"
-	"github.com/kai-scheduler/KAI-scheduler/test/e2e/modules/utils"
-	"github.com/kai-scheduler/KAI-scheduler/test/e2e/modules/wait"
-)
-
-const (
-	resizeTimeout = 2 * time.Minute
-	resizePoll    = 2 * time.Second
-)
-
-var _ = Describe("In-place pod resize", func() {
-	// setupTest must run inside each It: GetConnectivity captures the node's
-	// context, which ginkgo cancels when the node exits, so a TestContext
-	// created in BeforeEach is unusable by the time the It body runs.
-	setupTest := func(ctx context.Context) *testcontext.TestContext {
-		testCtx := testcontext.GetConnectivity(ctx, Default)
-		skipIfNoResizeSupport(testCtx)
-		capacity.SkipIfInsufficientClusterResources(testCtx.KubeClientset,
-			&capacity.ResourceList{
-				Cpu:      resource.MustParse("2"),
-				PodCount: 1,
-			},
-		)
-		return testCtx
-	}
-
-	It("rejects an upsize that exceeds the queue CPU limit", func(ctx context.Context) {
-		testCtx := setupTest(ctx)
-		defer testCtx.ClusterCleanup(ctx)
-
-		parentQueue := queue.CreateQueueObject(utils.GenerateRandomK8sName(10), "")
-		childQueue := queue.CreateQueueObject(utils.GenerateRandomK8sName(10), parentQueue.Name)
-		childQueue.Spec.Resources.CPU = v2.QueueResource{Quota: 500, OverQuotaWeight: 1, Limit: 2000}
-		testCtx.InitQueues([]*v2.Queue{childQueue, parentQueue})
-
-		pod := runPodWithCPU(ctx, testCtx, "500m")
-
-		err := resizeCPU(ctx, testCtx, pod, "3")
-		Expect(err).To(HaveOccurred(), "upsize past the queue limit should be denied")
-		Expect(err.Error()).To(ContainSubstring("resize rejected"))
-	})
-
-	It("allows upsizes within the queue limit and downsizes", func(ctx context.Context) {
-		testCtx := setupTest(ctx)
-		defer testCtx.ClusterCleanup(ctx)
-
-		parentQueue := queue.CreateQueueObject(utils.GenerateRandomK8sName(10), "")
-		childQueue := queue.CreateQueueObject(utils.GenerateRandomK8sName(10), parentQueue.Name)
-		childQueue.Spec.Resources.CPU = v2.QueueResource{Quota: 500, OverQuotaWeight: 1, Limit: 2000}
-		testCtx.InitQueues([]*v2.Queue{childQueue, parentQueue})
-
-		pod := runPodWithCPU(ctx, testCtx, "500m")
-
-		By("upsizing within the limit")
-		Expect(resizeCPU(ctx, testCtx, pod, "1")).To(Succeed())
-		expectEnactedCPU(ctx, testCtx, pod, "1")
-
-		By("downsizing")
-		Expect(resizeCPU(ctx, testCtx, pod, "250m")).To(Succeed())
-		expectEnactedCPU(ctx, testCtx, pod, "250m")
-	})
-
-	It("does not charge infeasible resize targets to the queue", func(ctx context.Context) {
-		testCtx := setupTest(ctx)
-		defer testCtx.ClusterCleanup(ctx)
-
-		parentQueue := queue.CreateQueueObject(utils.GenerateRandomK8sName(10), "")
-		childQueue := queue.CreateQueueObject(utils.GenerateRandomK8sName(10), parentQueue.Name)
-		testCtx.InitQueues([]*v2.Queue{childQueue, parentQueue})
-
-		pod := runPodWithCPU(ctx, testCtx, "500m")
-
-		By("requesting a resize no node can satisfy")
-		Expect(resizeCPU(ctx, testCtx, pod, "1000")).To(Succeed(),
-			"unbounded queue must not block the resize; feasibility is the kubelet's call")
-
-		By("waiting for the kubelet to mark the resize Infeasible")
-		Eventually(func(g Gomega) {
-			current := &v1.Pod{}
-			g.Expect(testCtx.ControllerClient.Get(ctx,
-				types.NamespacedName{Namespace: pod.Namespace, Name: pod.Name}, current)).To(Succeed())
-			g.Expect(isResizeInfeasible(current)).To(BeTrue())
-		}).WithContext(ctx).WithTimeout(resizeTimeout).WithPolling(resizePoll).Should(Succeed())
-
-		By("verifying the queue keeps charging the enacted request, not the infeasible spec")
-		Eventually(func(g Gomega) int64 {
-			q, err := testCtx.KubeAiSchedClientset.SchedulingV2().Queues("").
-				Get(ctx, childQueue.Name, metav1.GetOptions{})
-			g.Expect(err).NotTo(HaveOccurred())
-			allocated := q.Status.Allocated[v1.ResourceCPU]
-			return allocated.MilliValue()
-		}).WithContext(ctx).WithTimeout(resizeTimeout).WithPolling(resizePoll).
-			Should(Equal(int64(500)), "queue allocated must reflect the 500m enacted request")
-	})
-})
-
-// runPodWithCPU creates a pod in the test queue with the given CPU request and
-// waits for it to be running, so a ContainerStatus with enacted resources exists.
-func runPodWithCPU(ctx context.Context, testCtx *testcontext.TestContext, cpu string) *v1.Pod {
-	pod := rd.CreatePodObject(testCtx.Queues[0], v1.ResourceRequirements{
-		Requests: v1.ResourceList{
-			v1.ResourceCPU:    resource.MustParse(cpu),
-			v1.ResourceMemory: resource.MustParse("256Mi"),
-		},
-	})
-	pod, err := rd.CreatePod(ctx, testCtx.KubeClientset, pod)
-	Expect(err).NotTo(HaveOccurred(), "failed to create pod")
-	wait.ForPodReady(ctx, testCtx.ControllerClient, pod)
-	return pod
-}
-
-// resizeCPU updates the pod's CPU request through the pods/resize subresource.
-// Retries on conflict: the kubelet concurrently writes pod status during a
-// resize, so a Get-mutate-Update can race with it.
-func resizeCPU(ctx context.Context, testCtx *testcontext.TestContext, pod *v1.Pod, cpu string) error {
-	return retry.RetryOnConflict(retry.DefaultRetry, func() error {
-		current := &v1.Pod{}
-		if err := testCtx.ControllerClient.Get(ctx,
-			types.NamespacedName{Namespace: pod.Namespace, Name: pod.Name}, current); err != nil {
-			return err
-		}
-		current.Spec.Containers[0].Resources.Requests[v1.ResourceCPU] = resource.MustParse(cpu)
-		return testCtx.ControllerClient.SubResource("resize").Update(ctx, current)
-	})
-}
-
-func expectEnactedCPU(ctx context.Context, testCtx *testcontext.TestContext, pod *v1.Pod, cpu string) {
-	want := resource.MustParse(cpu)
-	Eventually(func(g Gomega) {
-		current := &v1.Pod{}
-		g.Expect(testCtx.ControllerClient.Get(ctx,
-			types.NamespacedName{Namespace: pod.Namespace, Name: pod.Name}, current)).To(Succeed())
-		g.Expect(current.Status.ContainerStatuses).NotTo(BeEmpty())
-		g.Expect(current.Status.ContainerStatuses[0].Resources).NotTo(BeNil())
-		got := current.Status.ContainerStatuses[0].Resources.Requests[v1.ResourceCPU]
-		g.Expect(got.Cmp(want)).To(BeZero(), "enacted cpu: want %s got %s", want.String(), got.String())
-	}).WithContext(ctx).WithTimeout(resizeTimeout).WithPolling(resizePoll).Should(Succeed())
-}
-
-func isResizeInfeasible(pod *v1.Pod) bool {
-	for _, c := range pod.Status.Conditions {
-		if c.Type == v1.PodResizePending {
-			return c.Status == v1.ConditionTrue && c.Reason == v1.PodReasonInfeasible
-		}
-	}
-	return false
-}
-
-// skipIfNoResizeSupport skips the suite on clusters older than 1.33, where
-// InPlacePodVerticalScaling is not enabled by default.
-func skipIfNoResizeSupport(testCtx *testcontext.TestContext) {
-	ver, err := testCtx.KubeClientset.Discovery().ServerVersion()
-	Expect(err).NotTo(HaveOccurred(), "failed to read server version")
-	minor, err := strconv.Atoi(strings.TrimSuffix(ver.Minor, "+"))
-	Expect(err).NotTo(HaveOccurred(), "failed to parse server minor version %q", ver.Minor)
-	if ver.Major != "1" || minor < 33 {
-		Skip("in-place pod resize requires Kubernetes 1.33+ (InPlacePodVerticalScaling on by default)")
-	}
-}
